Move Message List mocks
