+++ Tests/StreamChatUITests/Mocks/ChatMessageList/ChatMessageCell_Mock.swift
@@ -1,0 +1,17 @@
+//
+// Copyright © 2023 Stream.io Inc. All rights reserved.
+//
+
+@testable import StreamChat
+@testable import StreamChatUI
+import UIKit
+
+class ChatMessageCell_Mock: ChatMessageCell {
+    var mockedMessage: ChatMessage?
+
+    override var messageContentView: ChatMessageContentView? {
+        let view = ChatMessageContentView()
+        view.content = mockedMessage
+        return view
+    }
+}
