+++ Tests/StreamChatUITests/Mocks/ChatMessageList/ChatMessageListVCDataSource_Mock.swift
@@ -1,0 +1,33 @@
+//
+// Copyright © 2023 Stream.io Inc. All rights reserved.
+//
+
+@testable import StreamChat
+@testable import StreamChatUI
+import UIKit
+
+class ChatMessageListVCDataSource_Mock: ChatMessageListVCDataSource {
+    var mockedIsFirstPageLoaded: Bool = true
+    var isFirstPageLoaded: Bool {
+        mockedIsFirstPageLoaded
+    }
+
+    var mockedChannel: ChatChannel?
+    func channel(for vc: ChatMessageListVC) -> ChatChannel? {
+        mockedChannel
+    }
+
+    var messages: [StreamChat.ChatMessage] = []
+
+    func numberOfMessages(in vc: StreamChatUI.ChatMessageListVC) -> Int {
+        messages.count
+    }
+
+    func chatMessageListVC(_ vc: StreamChatUI.ChatMessageListVC, messageAt indexPath: IndexPath) -> StreamChat.ChatMessage? {
+        messages[indexPath.item]
+    }
+
+    func chatMessageListVC(_ vc: StreamChatUI.ChatMessageListVC, messageLayoutOptionsAt indexPath: IndexPath) -> StreamChatUI.ChatMessageLayoutOptions {
+        .init()
+    }
+}
