+++ Tests/StreamChatUITests/SnapshotTests/ChatMessageList/ChatMessageListVC_Tests.swift
@@ -640,140 +640,3 @@ final class ChatMessageListVC_Tests: XCTestCase {
         XCTAssertNotNil(cell.headerContainerView.subviews.first as? ChatMessageListDateSeparatorView)
     }
 }
-
-class ChatMessageListView_Mock: ChatMessageListView {
-    var mockIsLastCellFullyVisible = false
-    override var isLastCellFullyVisible: Bool {
-        mockIsLastCellFullyVisible
-    }
-
-    var reloadSkippedMessagesCallCount = 0
-    override func reloadSkippedMessages() {
-        reloadSkippedMessagesCallCount += 1
-    }
-
-    var reloadRowsCallCount = 0
-    var reloadRowsCalledWith: [IndexPath] = []
-    override func reloadRows(at indexPaths: [IndexPath], with animation: UITableView.RowAnimation) {
-        reloadRowsCallCount += 1
-        reloadRowsCalledWith += indexPaths
-    }
-
-    var mockedCellForRow: ChatMessageCell_Mock?
-    override func cellForRow(at indexPath: IndexPath) -> UITableViewCell? {
-        mockedCellForRow
-    }
-
-    var mockedIndexPathsForVisibleRows: [IndexPath]?
-    override var indexPathsForVisibleRows: [IndexPath]? {
-        mockedIndexPathsForVisibleRows ?? super.indexPathsForVisibleRows
-    }
-
-    var scrollToRowCallCount = 0
-    var scrollToRowCalledWith: IndexPath?
-    override func scrollToRow(at indexPath: IndexPath, at scrollPosition: UITableView.ScrollPosition, animated: Bool) {
-        scrollToRowCalledWith = indexPath
-        scrollToRowCallCount += 1
-    }
-
-    var mockMoreContentThanOnePage = false {
-        didSet {
-            if mockMoreContentThanOnePage {
-                contentSize.height = 100
-                bounds = .init(origin: .zero, size: .init(width: 0, height: 10))
-            } else {
-                contentSize.height = 10
-                bounds = .init(origin: .zero, size: .init(width: 0, height: 100))
-            }
-        }
-    }
-
-    var scrollToMostRecentMessageCallCount = 0
-    override func scrollToMostRecentMessage(animated: Bool = true) {
-        scrollToMostRecentMessageCallCount += 1
-    }
-
-    var updateMessagesCompletion: (() -> Void)?
-    override func updateMessages(with changes: [ListChange<ChatMessage>], completion: (() -> Void)? = nil) {
-        updateMessagesCompletion = completion
-    }
-}
-
-class ChatMessageCell_Mock: ChatMessageCell {
-    var mockedMessage: ChatMessage?
-
-    override var messageContentView: ChatMessageContentView? {
-        let view = ChatMessageContentView()
-        view.content = mockedMessage
-        return view
-    }
-}
-
-class ChatMessageListVCDataSource_Mock: ChatMessageListVCDataSource {
-    var mockedIsFirstPageLoaded: Bool = true
-    var isFirstPageLoaded: Bool {
-        mockedIsFirstPageLoaded
-    }
-
-    var mockedChannel: ChatChannel?
-    func channel(for vc: ChatMessageListVC) -> ChatChannel? {
-        mockedChannel
-    }
-
-    var messages: [StreamChat.ChatMessage] = []
-
-    func numberOfMessages(in vc: StreamChatUI.ChatMessageListVC) -> Int {
-        messages.count
-    }
-
-    func chatMessageListVC(_ vc: StreamChatUI.ChatMessageListVC, messageAt indexPath: IndexPath) -> StreamChat.ChatMessage? {
-        messages[indexPath.item]
-    }
-
-    func chatMessageListVC(_ vc: StreamChatUI.ChatMessageListVC, messageLayoutOptionsAt indexPath: IndexPath) -> StreamChatUI.ChatMessageLayoutOptions {
-        .init()
-    }
-}
-
-class ChatMessageListVCDelegate_Mock: ChatMessageListVCDelegate {
-    var mockedHeaderView: ChatMessageDecorationView?
-    var mockedFooterView: ChatMessageDecorationView?
-
-    func chatMessageListVC(_ vc: ChatMessageListVC, willDisplayMessageAt indexPath: IndexPath) {}
-
-    func chatMessageListVC(_ vc: ChatMessageListVC, scrollViewDidScroll scrollView: UIScrollView) {}
-
-    func chatMessageListVC(_ vc: ChatMessageListVC, didTapOnAction actionItem: ChatMessageActionItem, for message: ChatMessage) {}
-
-    func chatMessageListVC(_ vc: ChatMessageListVC, didTapOnMessageListView messageListView: ChatMessageListView, with gestureRecognizer: UITapGestureRecognizer) {}
-
-    var shouldLoadPageAroundMessageCallCount = 0
-    var shouldLoadPageAroundMessageResult: Error?
-    func chatMessageListVC(_ vc: ChatMessageListVC, shouldLoadPageAroundMessage message: ChatMessage, _ completion: @escaping ((Error?) -> Void)) {
-        shouldLoadPageAroundMessageCallCount += 1
-        if let result = shouldLoadPageAroundMessageResult {
-            completion(result)
-        }
-    }
-
-    var shouldLoadFirstPageCallCount = 0
-    func chatMessageListVCShouldLoadFirstPage(_ vc: ChatMessageListVC) {
-        shouldLoadFirstPageCallCount += 1
-    }
-
-    func chatMessageListVC(
-        _ vc: ChatMessageListVC,
-        headerViewForMessage message: ChatMessage,
-        at indexPath: IndexPath
-    ) -> ChatMessageDecorationView? {
-        mockedHeaderView
-    }
-
-    func chatMessageListVC(
-        _ vc: ChatMessageListVC,
-        footerViewForMessage message: ChatMessage,
-        at indexPath: IndexPath
-    ) -> ChatMessageDecorationView? {
-        mockedFooterView
-    }
-}
